Move some functions around; add test/fixture for 'remarkdown' command.
