+++ index.js
@@ -1,0 +1,65 @@
+import { unified } from 'unified';
+import rehypeParse from 'rehype-parse';
+import rehypeRaw from 'rehype-raw';
+import rehypeRemark from 'rehype-remark';
+import rehypeSanitize from 'rehype-sanitize';
+import rehypeStringify from 'rehype-stringify';
+import remarkParse from 'remark-parse';
+import remarkRehype from 'remark-rehype';
+import remarkStringify from 'remark-stringify';
+import remarkGfm from 'remark-gfm';
+import { SKIP, visit } from 'unist-util-visit';
+import { fromHtml } from 'hast-util-from-html';
+import { toMdast } from 'hast-util-to-mdast';
+
+export async function slurp(stream) {
+	let arr = [],
+		len = 0;
+	for await (let chunk of stream) {
+		arr.push(chunk);
+		len += chunk.length;
+	}
+	return Buffer.concat(arr, len).toString();
+}
+
+/*
+	Remark plugin to convert mdast 'html' nodes
+	into mdast Markdown nodes.
+*/
+function remarkParseHtml() {
+	return function (tree, file) {
+		visit(tree, 'html', (node, index, parent) => {
+			const mdast = toMdast(fromHtml(node, { fragment: true }), {
+				document: false
+			});
+			parent.children.splice(index, 1, ...mdast.children);
+			return [SKIP, index];
+		});
+	};
+}
+
+export function getHtmlToMdProcessor(opts = {}) {
+	return unified()
+		.use(rehypeParse)
+		.use(rehypeRemark)
+		.use(remarkGfm)
+		.use(remarkStringify, opts);
+}
+
+export function getMdToMdProcessor(opts = {}) {
+	return unified()
+		.use(remarkParse)
+		.use(remarkGfm)
+		.use(remarkParseHtml)
+		.use(remarkStringify, opts);
+}
+
+export function getMdToHtmlProcessor(opts) {
+	return unified()
+		.use(remarkParse)
+		.use(remarkGfm)
+		.use(remarkRehype, { allowDangerousHtml: true })
+		.use(rehypeRaw)
+		.use(rehypeSanitize)
+		.use(rehypeStringify);
+}
