+++ package.json
@@ -2,9 +2,10 @@
 	"name": "trimd",
 	"version": "0.3.1",
 	"description": "Trim HTML to Markdown from the command line.",
-	"main": "cli.js",
+	"main": "index.js",
 	"files": [
-		"cli.js"
+		"cli.js",
+		"index.js"
 	],
 	"bin": {
 		"trimd": "cli.js"
@@ -29,7 +30,8 @@
 	"devDependencies": {
 		"esbuild": "^0.17.10",
 		"prettier": "^2.8.4",
-		"pretty-quick": "^3.1.3"
+		"pretty-quick": "^3.1.3",
+		"tape": "^5.6.3"
 	},
 	"dependencies": {
 		"hast-util-from-html": "^1.0.1",
