+++ test/fixtures/mixed.md
@@ -1,0 +1,3 @@
+<p>This file contains both HTML, including things like <b>bold</b> <a href="https://joinmastodon.org">and a link</a></p>
+
+> It also contains _markdown_, most notably [links](https://joinmastodon.org) in .md format
