+++ test/fixtures/mixed.result.md
@@ -1,0 +1,3 @@
+This file contains both HTML, including things like **bold** [and a link](https://joinmastodon.org)
+
+> It also contains _markdown_, most notably [links](https://joinmastodon.org) in .md format
